重构: 同意导入路径风格，重构HLSL常量代码
1 移除导入路径中的 .js 后缀，统一模块导入风格
2 抽离 HLSL 常量至 scripts/shared.HlslData，新增 hlsl.utils.ts 提供 isHlslKeyword
